content/html/content/test/file_iframe_sandbox_c_if2.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 341604</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

</head>
<script type="text/javascript">
  function ok(result, desc) {
    window.parent.postMessage({ok: result, desc: message}, "*");
  }

  function doStuff() {
    ok(false, "documents sandboxed without allow-scripts should NOT be able to run inline scripts");
  }
</script>
<script src='file_iframe_sandbox_fail.js'></script>
<body onLoad='window.parent.postmessage({ok: false, desc: "documents sandboxed without allow-scripts should NOT be able to run script from event handlers"}, "*");doStuff();'>
  I am sandboxed with no permissions
  <img src="about:blank" onerror='ok(false, "documents sandboxed without allow-scripts should NOT be able to run script from event handlers");')>
</body>
</html>
